feat: vastly improve scripts and implement configs

+++ dotfiles/bin/.local/scripts/start-remote-gdm3.sh
@@ -1,5 +1,8 @@
 #!/bin/bash
 
+# NOTE: Script doesn't work as-is - needs more work, wayland likes getting in the way
+echo "Script doesn't work as-is - needs more work, wayland likes getting in the way"
+
 function prev_cmd_successful() {
   if [ $? -ne 0 ]; then
     echo 'Failure in process. Exiting and cleaning up.'
@@ -34,21 +37,49 @@ if [[ "$key" != "y" ]]; then
   exit
 fi
 
-echo 'Backing up previous gdm/custom.conf'
-sudo cp /etc/gdm3/custom.conf /etc/gdm3/custom.conf.bak
 
 # TODO: change to be better - don't hard code username (in flyby file)
 # use script and replacement string to do this.
-sudo cp $HOME/common-setups/dotfiles/Z_NO_STOW/ZZZZ_FLYBYS/gdm3/custom.autologin.conf /etc/gdm3/custom.conf
 
+mgmt_file="/etc/gdm3/bash-managed.txt"
+if [[ ! -f "$mgmt_file" ]]; then
+  echo 'Backing up previous gdm/custom.conf'
+  sudo cp /etc/gdm3/custom.conf /etc/gdm3/custom.conf.bak
+  echo 'Setting up configs.'
+  sudo cp $HOME/common-setups/dotfiles/Z_NO_STOW/ZZZZ_FLYBYS/gdm3/custom.autologin.conf /etc/gdm3/custom.conf
+  sudo touch "$mgmt_file"
+else
+  echo 'Management file detected - NOOP!'
+fi
+
+stop-remote-gdm3.sh "keep-conf"
 sudo systemctl start gdm3
-prev_cmd_successful
-sleep 12
-prev_cmd_successful
 
-# DISPLAY=:0 XAUTHORITY=/run/user/1000/gdm/Xauthority rustdesk &
+echo "Waiting for X11 session..."
 
-# OR:
+loop_limit=30
+loop_iteration=0
+
+current_sessions=$(loginctl list-sessions --no-legend | awk "/$USER/"'{print $1}')
+
+while [[ $loop_iteration -le $loop_limit ]]; do
+
+  while IFS= read -r SESSION_ID; do
+
+    if [ -n "$SESSION_ID" ]; then
+      TYPE=$(loginctl show-session "$SESSION_ID" -p Type --value)
+      DISPLAY=$(loginctl show-session "$SESSION_ID" -p Display --value)
+
+      if [ "$TYPE" = "x11" ] && [ -n "$DISPLAY" ]; then
+        echo "X11 session ready on $DISPLAY"
+        break
+      fi
+    fi
+  done <<< "$current_sessions"
+
+  ((loop_iteration++))
+  sleep 1
+done
 
 # TODO: change to be better - don't hard code username
 sudo -u cragady DISPLAY=:0 XAUTHORITY=$(either_or_file) rustdesk &
